--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/DelegateWrappersInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/DelegateWrappersInternal.cs
@@ -326,8 +326,8 @@ namespace Proto.Promises
                 {
                     handler.SuppressRejection = true;
                     owner._rejectContainer = handler._rejectContainer;
-                    // Very important, write State must come after write _result and _valueContainer. This is a volatile write, so we don't need a full memory barrier.
-                    // State is checked for completion, and if it is read not pending on another thread, _result and _valueContainer must have already been written so the other thread can read them.
+                    // Very important, write State must come after write _result and _rejectContainer. This is a volatile write, so we don't need a full memory barrier.
+                    // State is checked for completion, and if it is read not pending on another thread, _result and _rejectContainer must have already been written so the other thread can read them.
                     owner.State = handler.State;
                     handler.MaybeDispose();
                     handler = owner;