Master:

|         Type | Pending |       Mean |    Error |   StdDev | Allocated | Survived |
|------------- |-------- |-----------:|---------:|---------:|----------:|---------:|
|   AsyncAwait |   False |   591.4 ns |  2.77 ns |  2.59 ns |         - |        - |
| ContinueWith |   False |   869.6 ns |  5.67 ns |  4.73 ns |         - |        - |
|   AsyncAwait |    True | 2,948.9 ns | 12.30 ns | 11.51 ns |         - |    688 B |
| ContinueWith |    True | 3,238.9 ns | 11.37 ns |  9.49 ns |         - |    416 B |

This PR:

|         Type | Pending |       Mean |    Error |   StdDev | Allocated | Survived |
|------------- |-------- |-----------:|---------:|---------:|----------:|---------:|
|   AsyncAwait |   False |   338.3 ns |  1.79 ns |  1.59 ns |         - |        - |
| ContinueWith |   False |   518.3 ns |  1.19 ns |  1.11 ns |         - |        - |
|   AsyncAwait |    True | 2,270.0 ns | 16.88 ns | 15.79 ns |         - |    688 B |
| ContinueWith |    True | 2,827.3 ns | 16.96 ns | 15.04 ns |         - |    416 B |